Fix edition locking timer so it always expires
The timer implementation could get into a state where it never expired, which doesn't really make sense in the edition locking context. In the previous commit we worked around this by sending another lease request for the same edition if we didn't get a success response for the lease request, but that isn't how users might behave in reality. However, we can guarantee that Whitehall will mark the leases as timed out after a certain amount of time, so the new model is more true to reality.